requirements.txt
apache-airflow[celery,postgres,redis,SQLAlchemy]==2.3.3
